Simulation/requirements.txt
eclypse==0.6.19
networkx==3.4.2
PyYAML==6.0.2
pyswip==0.3.2
matplotlib==3.10.3
seaborn==0.13.2